Add binary that counts CPU cycles for various types of loops structures.
This new example binary implements a small benchmarking framework that
allows us to measure the number of CPU cycles that spent executing
a given piece of code.

This in turn allows us to try and gauge the number of CPU cycles taken
up by specific instructions.  The `cycles` module encodes our best
guesses for the number of cycles per instruction, under various
circumstances. It's possible some of these guesses are wrong, and
I clearly don't have a consistent model for the CPU behavior since
various instructions have multiple possible CPU cycle values, even for
what otherwise seem to be the same situations.

See
https://ctrlsrc.io/posts/2023/counting-cpu-cycles-on-esp32c3-esp32c6/,
which discusses this approach a bit more.