Initial `niccle_esp::eth_phy` module: periodic LTPs via interrupts.
Transmitting LTPs takes very little time, so we do so directly from
within the interrupt handler. This initial scaffolding will be expanded
in a subsequent commit to support transmission of data packets from the
main thread.

Note that the code running in the interrupt handler needs to be able to
access various resources such as the timer instance and the `TxStats`
instance, while the code in the main thread of execution also needs
access to those resources (e.g. to read and print the `TxStats` values).
To play nice with the type system/borrow checker we therefore model
these as resources that will be shared between both threads of
execution, and we therefore protect them with a `Mutex<RefCell<...>>`.
All of this is encapsulated behind the public `eth_phy::Phy` and
`eth_phy::InterruptHandler` interfaces.

The use of shared resources will likely expand in subsequent commits.
E.g. the TX pin (or rather, the CPU output signal connected to that pin)
will need to be used from the main thread when transmitting data
packets, and hence is already modeled as a shared resource as well.

It's worth noting that `'static` lifetimes are used in a few of the
generic types. In most cases these could be relaxed to non-static named
lifetimes, but for the time being I'm not using the library in any way
that requires more relaxed lifetimes. Using `'static` means we need to
propagate a few less named lifetimes across all of the different structs
and their impls, keeping things a bit simpler.

This commit also defines a new, currently-unfinished example binary
which we'll use to test the transmission functionality with. It
currently invokes the library code to transmit periodic LTPs, but will
be expanded to include the transmission of a simple data packet in
a subsequent commit.