Expand `niccle_esp::eth_phy` to support transmitting packets.
To ensure accurate timing the transmission loop is implemented directly
in assembly, using the dedicated IO feature of the ESP32-C6 chip.

This implementation does still have a small issue: the transmission
timing of the first byte of data is off by 6.25ns, since during the
first iteration of the loop the `bne` instruction takes 2 cycles, rather
than the 1 cycle it takes in the subsequent iterations. That will be
fixed in a later commit.

A hardcoded test packet (an ARP request) has been added to the
`niccle::example_data` module, and this packet is used in the
`arp_request` example binary to pass to the PHY layer. When running the
binary I'm able to see the incoming packets using Wireshark on my
computer.