Implement RX support for `niccle_esp::eth_phy`, plus a basic MAC layer.
Receipt of incoming Ethernet packets is handled via an edge-triggered
interrupt, which in turn calls a function which implements a receive
loop in assembly code using the dedicated IO feature of the ESP32-C6
chip.

The receive loop tries to align its sampling period with a falling edge
in the preamble, and subsequently samples every 50ns, one sample per bit
symbol half. It performs Manchester decoding as part of the loop code.

In order for this to work, we must use the `direct-vectoring` feature of
the ESP32-C6 hal crate. Without it there is too much latency (about
7500ns) between the interrupt being triggered and the interrupt routine
being invoked, which means that the receive loop doesn't have a chance
of observing the signal preamble as a result. With the direct vectoring
feature the interrupt routine is invoked within about 844ns of the first
signal edge, at which point there is still more than 5500ns of preamble
signal left to observe.

In order to test that the receive loop actually works, we need to know
whether the data it returns is actually valid. Hence, this commit
includes the very basic beginning of a MAC layer. This code currently
processes the packet data by finding the SFD byte, aligning the frame
data, and then calculating the checksum and comparing it with the
received FCS bytes. It only logs some diagnostics, and doesn't do
anything else with the received packets. A subsequent change will expand
this functionality.

See https://ctrlsrc.io/posts/2023/niccle-receiving-data-at-the-phy-layer
for more info.