Implement a MAC layer.
The MAC layer handles prepending the preamble and SFD to outgoing frames
before passing them to the PHY, as well as frame padding and FCS
generation. It handles stripping of preamble and SFD from incoming
packets, and verification of the frame data using the received FCS.

The MAC RX implementation will buffer a preconfigured amount of incoming
frames before it starts dropping incoming packets, using the `thingbuf`
as the backing data structure. It provides a convenient lock-free queue
interface allowing us to push and pop fixed-sized frame buffers at a
time without needing to use a mutex to protect it.

In the future the implementation could improved further by avoiding a
few copies that are unnecessary, or by using something like a ring buffer
to avoid having to allocate a large fixed-size frame buffer to each
frame, even those frames are much smaller than a full frame.

This change also reduces the log levels of some of the logs so that it's
feasible to run with `Debug`-level logs without spamming the output.