Add a `smoltcp_http_get` example binary which issues a GET request.
Most of this code is based on the `httpclient.rs` example in the SmolTCP
repo:
https://github.com/smoltcp-rs/smoltcp/blob/v0.11.0/examples/httpclient.rs.

A testdata/ directory with files of various sizes containing random
content is also included, together with some scripts with which the data
can be served over HTTP (ensuring the response is deterministic each
time) and with which the expected CRC32 values for each response for
each file can be determined, to be compared against the output of the
binary.

A pcaps/ directory is also included with a sample wireshark capture and
corresponding serial output from the `smoltcp_http_get` binary. The
wireshark capture shows many packet retransmissions for each connection,
while the serial output shows a high rate of invalid packets (at least
30% of all the received packets, which is likely an underestimate).